+++ Prog2 Practis6/ArticleBook.cs
@@ -9,7 +9,7 @@ namespace Prog2_Practis6
 
         private string _publisher;
 
-        public ArticleBook(string title, string author, string publisher, uint article, uint price) : base(article, price)
+        public ArticleBook(string title, string author, string publisher, uint article, double price) : base(article, price)
         {
             _title = title;
             _author = author;
